requirements.txt
pandas
openpyxl
sqlalchemy
psycopg2-binary
python-dotenv
google-api-python-client
google-auth